refactor: TT-328 Enum들을 패키지에 모음
