+++ src/main/java/com/twentythree/peech/user/dto/request/CreateUserRequestDTO.java
@@ -1,9 +1,9 @@
 package com.twentythree.peech.user.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
-import com.twentythree.peech.user.AuthorizationServer;
-import com.twentythree.peech.user.UserGender;
-import com.twentythree.peech.user.UserRole;
+import com.twentythree.peech.user.value.AuthorizationServer;
+import com.twentythree.peech.user.value.UserGender;
+import com.twentythree.peech.user.value.UserRole;
 import lombok.Data;
 
 import java.time.LocalDate;
