+++ src/main/java/com/twentythree/peech/user/validator/UserValidator.java
@@ -1,6 +1,6 @@
 package com.twentythree.peech.user.validator;
 
-import com.twentythree.peech.user.AuthorizationIdentifier;
+import com.twentythree.peech.user.entity.AuthorizationIdentifier;
 
 public interface UserValidator {
     boolean NickNameIsNotDuplicated(String nickName);
