+++ src/main/java/com/twentythree/peech/user/validator/UserValidatorImpl.java
@@ -1,7 +1,6 @@
 package com.twentythree.peech.user.validator;
 
-import com.twentythree.peech.user.AuthorizationIdentifier;
-import com.twentythree.peech.user.AuthorizationServer;
+import com.twentythree.peech.user.entity.AuthorizationIdentifier;
 import com.twentythree.peech.user.entity.UserEntity;
 import com.twentythree.peech.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
