+++ src/main/java/com/twentythree/peech/user/value/AuthorizationServer.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.user;
+package com.twentythree.peech.user.value;
 
 public enum AuthorizationServer {
     KAKAO, APPLE
