+++ src/main/java/com/twentythree/peech/user/value/SignUpFinished.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.user;
+package com.twentythree.peech.user.value;
 
 public enum SignUpFinished {
     PENDING, FINISHED
