+++ src/main/java/com/twentythree/peech/user/value/UserGender.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.user;
+package com.twentythree.peech.user.value;
 
 public enum UserGender {
     MALE, FEMALE
