+++ src/main/java/com/twentythree/peech/user/value/UserRole.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.user;
+package com.twentythree.peech.user.value;
 
 public enum UserRole {
     ROLE_COMMON, ROLE_ADMIN
