+++ src/main/java/com/twentythree/peech/user/value/UserStatus.java
@@ -1,4 +1,4 @@
-package com.twentythree.peech.user;
+package com.twentythree.peech.user.value;
 
 public enum UserStatus {
     ACTIVE, DELETE
